Search result titles include duration and fallback to object id when title isn't present The commit restores the pre-BL8 title behavior by moving the search_result_label helper into the component and splitting some time formatting helpers into their own helper file to make a more minimal include into the component.  Also setup testing for components.
