+++ app/components/index_header_media_object_component.rb
@@ -1,7 +1,10 @@
 class IndexHeaderMediaObjectComponent < Blacklight::DocumentTitleComponent
+  include TimeFormattingHelper
+
   def initialize(title = nil, document: nil, presenter: nil, as: :h3, counter: nil, classes: 'index_title document-title-heading col', link_to_document: true, document_component: nil, actions: true)
     super
     @classes += @actions.present? ? " col-sm-9 col-lg-10" : " col-md-12"
+    @title = search_result_label(presenter.document)
   end
 
   # Override to add test-id
@@ -11,5 +14,24 @@ class IndexHeaderMediaObjectComponent < Blacklight::DocumentTitleComponent
     else
       content_tag('span', @title.presence || content.presence || presenter.heading, itemprop: 'name')
     end
+  end
+
+  private
+
+  def search_result_label document
+    if document['title_tesi'].present?
+      label = truncate(document['title_tesi'], length: 100)
+    else
+      label = document[:id]
+    end
+
+    if document['duration_ssi'].present?
+      duration = document['duration_ssi']
+      if duration.respond_to?(:to_i) && duration.to_i > 0
+        label += " (#{milliseconds_to_formatted_time(duration.to_i, false)})"
+      end
+    end
+
+    label
   end
 end
